==> data/5-codebook.Rmd <==
---
title: "Codebook for processed data"
subtitle: "Add data labels, metadata, codebook, distributions, and convert to other file types"
author: "Ian Hussey"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    code_folding: hide
    highlight: haddock
    theme: flatly
    toc: yes
    toc_float: yes
---

```{r include=FALSE}
knitr::opts_chunk$set(message=FALSE,
                      warning=FALSE)
```

# Codebook

```{r}

# Following codebook package's [vignette](https://cran.r-project.org/web/packages/codebook/vignettes/codebook_tutorial.html)

# Dependencies

library(tidyverse)
library(knitr)
library(kableExtra)
library(codebook)
library(labelled)
library(rio)


# Add labels ----

# read data
data_codebook <- read.csv("processed/4_data_participant_level_with_hand_scoring.csv") 

# read data data_dictionary
data_dictionary <- read.csv("processed/5_data_dictionary.csv")

# add data_dictionary as labels
var_label(data_codebook) <- data_dictionary %>% 
  select(variable, label) %>% 
  dict_to_list()


# Add meta data ----

metadata(data_codebook)$name <- "Evaluative learning via deepfaked media"
metadata(data_codebook)$description <- "Across multiple experiments, we demonstrated that 'deepfakes' can establish automatic biases, self-reported evaluations, and behavioural intentions."
#metadata(data_codebook)$identifier <- "https://dx.doi.org/XXXXXXX"

metadata(data_codebook)$creator <- "Sean Hughes"
metadata(data_codebook)$citation <- "Hughes, S., Fried, O., Ferguson, M. J., Yao, D., Hughes, C., Hughes, R., & Hussey, I. (2020). Using Deepfakes to Hack the Human Mind."
metadata(data_codebook)$url <- "https://github.com/Sean-Hughes/DF-Impression-Formation--Video-and-Audio-"

# other meta data:  see https://schema.org/Dataset
metadata(data_codebook)$datePublished <- "2020"
metadata(data_codebook)$spatialCoverage <- "Online" 


# Create codebook ----

data_codebook %>%
  select(-behavioral_intentions_share, -behavioral_intentions_subscribe, -behavioral_intentions_recommend) %>%
  codebook()

```

- dropped item level behavioural intentions from codebook as they throw an error I couldn't solve.

# Write alternative file types to disk

Original csv file is used for analyses (as it is simplest), but other file types that integrate the labels are likely to be more useful for reuse. 

I include an R .rds file (which includes data labels and data types), SPSS .sav and Stata .dta files. 

```{r}

write_rds(data_codebook, "processed/4_data_participant_level_with_hand_scoring.rds") 
rio::export(data_codebook, "processed/4_data_participant_level_with_hand_scoring.sav") # SPSS file
rio::export(data_codebook, "processed/4_data_participant_level_with_hand_scoring.dta") # Stata file

```

# Session Info

```{r}

sessionInfo()

```
